002-get-movie-detail restTemplate 이용 정보 가져오기 수정
