+++ movie-api/src/main/java/com/zerobase/domain/response/movie/detail/ActorDto.java
@@ -1,4 +1,4 @@
-package com.zerobase.domain;
+package com.zerobase.domain.response.movie.detail;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +7,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
-public class StaffDto {
+public class ActorDto {
     private String peopleNm;
     private String peopleNmEn;
-    private String staffRoleNm;
+    private String cast;
+    private String castEn;
 }
