+++ movie-api/src/main/java/com/zerobase/domain/response/movie/detail/CompanyDetailDto.java
@@ -1,4 +1,4 @@
-package com.zerobase.domain;
+package com.zerobase.domain.response.movie.detail;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,9 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class CompanyDto {
+public class CompanyDetailDto {
     private String companyCd;
     private String companyNm;
+    private String companyNmEn;
+    private String companyPartNm;
 }
