+++ movie-api/src/main/java/com/zerobase/domain/response/movie/detail/GenreDto.java
@@ -1,4 +1,4 @@
-package com.zerobase.domain;
+package com.zerobase.domain.response.movie.detail;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
