+++ movie-api/src/main/java/com/zerobase/domain/response/movie/detail/MovieInfo.java
@@ -1,0 +1,12 @@
+package com.zerobase.domain.response.movie.detail;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class MovieInfo {
+    private MovieDetailDto movieInfo;
+}
