+++ movie-api/src/main/java/com/zerobase/domain/response/movie/detail/MovieInfoResult.java
@@ -1,4 +1,4 @@
-package com.zerobase.domain;
+package com.zerobase.domain.response.movie.detail;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class NationDto {
-    private String nationNm;
+public class MovieInfoResult {
+    private MovieDetailDto movieInfo;
 }
