+++ movie-api/src/main/java/com/zerobase/domain/response/movie/list/CompanyDto.java
@@ -1,4 +1,4 @@
-package com.zerobase.domain;
+package com.zerobase.domain.response.movie.list;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,9 +7,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class CompanyDetailDto {
+public class CompanyDto {
     private String companyCd;
     private String companyNm;
-    private String companyNmEn;
-    private String companyPartNm;
 }
