+++ movie-api/src/main/java/com/zerobase/domain/response/movie/list/DirectorDto.java
@@ -1,12 +1,14 @@
-package com.zerobase.domain.response;
+package com.zerobase.domain.response.movie.list;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
+@Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class Root {
-    private MovieListResult movieListResult;
+public class DirectorDto {
+    private String peopleNm;
 }
