+++ movie-api/src/main/java/com/zerobase/domain/response/movie/list/MovieListResult.java
@@ -1,4 +1,4 @@
-package com.zerobase.domain.response;
+package com.zerobase.domain.response.movie.list;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
