+++ movie-api/src/main/java/com/zerobase/domain/response/movie/list/Root.java
@@ -1,4 +1,4 @@
-package com.zerobase.domain;
+package com.zerobase.domain.response.movie.list;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,6 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class DirectorDetailDto {
-    private String peopleNm;
-    private String peopleNmEn;
+public class Root {
+    private MovieListResult movieListResult;
 }
