+++ movie-api/src/main/resources/application.properties
@@ -12,3 +12,7 @@ spring.data.redis.port=6380
 
 spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto=update
+
+API_BASE_URL=http://www.kobis.or.kr/kobisopenapi/webservice/rest/movie/
+API_MOVIES_URL=searchMovieList.json?key=
+API_MOVIE_DETAIL_URL=searchMovieInfo.json?key=
